+++ tests/testthat/test-split.r
@@ -80,3 +80,16 @@ test_that("str_split can split sentences correctly", {
    list(c("This is a sentence. ", "Is this a sentence? ", "Why, yes it is."))
   )
 })
+
+test_that("str_split can split sentences correctly", {
+  test <- "This is a sentence. Is this a sentence? Why, yes it is."
+
+  expect_that(
+    length(str_split(test, boundary("sentence"))[[1]]),
+    equals(3))
+  expect_that(
+    str_split(test, boundary("sentence")),
+    equals(list(c("This is a sentence. ", "Is this a sentence? ",
+                  "Why, yes it is."))))
+
+})