+++ R/trunc.R
@@ -18,9 +18,7 @@ str_trunc <- function(string, width, side = c("right", "left", "center"),
                       ellipsis = "...") {
   side <- match.arg(side)
 
-  too_long <- str_length(string) > width
-  too_long[is.na(too_long)] <- FALSE
-
+  too_long <- !is.na(string) && str_length(string) > width
   width... <- width - str_length(ellipsis)
 
   string[too_long] <- switch(side,