+++ tests/testthat/test-replace.r
@@ -48,6 +48,14 @@ test_that("replacement must be a string", {
   expect_equal(str_replace("xyz", "x", NA_character_), NA_character_)
 })
 
+test_that("can replace all types of NA values", {
+  expect_equal(str_replace_na(NA), "NA")
+  expect_equal(str_replace_na(NA_character_), "NA")
+  expect_equal(str_replace_na(NA_complex_), "NA")
+  expect_equal(str_replace_na(NA_integer_), "NA")
+  expect_equal(str_replace_na(NA_real_), "NA")
+})
+
 
 # functions ---------------------------------------------------------------