+++ tests/testthat/test-trim.r
@@ -16,3 +16,9 @@ test_that("side argument restricts trimming", {
   expect_equal(str_trim(" abc ", "left"),  "abc ")
   expect_equal(str_trim(" abc ", "right"), " abc")
 })
+
+test_that("str_squish removes excess spaces from all parts of string", {
+  expect_equal(str_squish("ab\t\tc\t"),   "ab c")
+  expect_equal(str_squish("\ta  bc"),   "a bc")
+  expect_equal(str_squish("\ta\t bc\t"), "a bc")
+})