+++ R/trim.R
@@ -17,3 +17,14 @@ str_trim <- function(string, side = c("both", "left", "right")) {
     both =  stri_trim_both(string)
   )
 }
+
+#' Trim whitespace from start, middle, and end of string.
+#'
+#' @export
+#' @rdname str_trim
+#' @examples
+#' str_squish("  String with trailing,  middle, and leading white space\t")
+#' str_squish("\n\nString with excess,  trailing and leading white   space\n\n")
+str_squish <- function(string) {
+  stri_trim_both(str_replace_all(string,"\\s+"," "))
+}