+++ R/trim.R
@@ -1,4 +1,7 @@
-#' Trim whitespace from start and end of string.
+#' Trim whitespace from a string
+#'
+#' `str_trim()` removes whitespace from start and end of string; `str_squish()`
+#' also reduces repeated whitespace inside a string.
 #'
 #' @param string A character vector.
 #' @param side Side on which to remove whitespace (left, right or both).
@@ -8,6 +11,9 @@
 #' @examples
 #' str_trim("  String with trailing and leading white space\t")
 #' str_trim("\n\nString with trailing and leading white space\n\n")
+#'
+#' str_squish("  String with trailing,  middle, and leading white space\t")
+#' str_squish("\n\nString with excess,  trailing and leading white   space\n\n")
 str_trim <- function(string, side = c("both", "left", "right")) {
   side <- match.arg(side)
 
@@ -18,13 +24,8 @@ str_trim <- function(string, side = c("both", "left", "right")) {
   )
 }
 
-#' Trim whitespace from start, middle, and end of string.
-#'
 #' @export
 #' @rdname str_trim
-#' @examples
-#' str_squish("  String with trailing,  middle, and leading white space\t")
-#' str_squish("\n\nString with excess,  trailing and leading white   space\n\n")
 str_squish <- function(string) {
   stri_trim_both(str_replace_all(string,"\\s+"," "))
 }