+++ NEWS.md
@@ -1,5 +1,9 @@
 # stringr 1.2.0.9000
 
+* New `str_glue()` and `str_glue_data()` provide convenient wrappers around
+  `glue` and `glue_data()` from the [glue](http://glue.tidyverse.org/) package
+  (#157).
+
 * Long deprecated `str_join()`, `ignore.case()` and `perl()` have now been 
   removed.