+++ NEWS.md
@@ -12,10 +12,15 @@
 * `str_remove()` and `str_remove_all()` functions. These wrap 
   `str_replace()` and `str_replace_all()` to remove patterns from strings.
   (@Shians, #178)
+  
+* `str_squish()` removes spaces from both the left and right side of strings, and also converts multiple space (or space-like characters) to a single space within strings. (@stephlocke #197)
 
 ## Bug fixes and minor improvements
 
 * `str_trunc()` now preserves NAs (@ClaytonJY, #162)
+
+* `str_trunc()` now throws an error when `width` is shorter than `ellipsis`
+  (@ClaytonJY, #163).
 
 * Long deprecated `str_join()`, `ignore.case()` and `perl()` have now been 
   removed.